fix test failures due to state leakage
