+++ extensions/amp-ad-network-cloudflare-impl/0.1/test/test-amp-ad-network-cloudflare-impl.js
@@ -18,10 +18,10 @@ import {AmpAdNetworkCloudflareImpl} from '../amp-ad-network-cloudflare-impl';
 import {
   AmpAdXOriginIframeHandler, // eslint-disable-line no-unused-vars
 } from '../../../amp-ad/0.1/amp-ad-xorigin-iframe-handler';
-import * as sinon from 'sinon';
 import {cloudflareIsA4AEnabled} from '../cloudflare-a4a-config';
 import {createElementWithAttributes} from '../../../../src/dom';
 import {createIframePromise} from '../../../../testing/iframe';
+import {initializeElement} from '../../../amp-a4a/0.1/test/utils';
 import * as vendors from '../vendors';
 
 describe('cloudflare-a4a-config', () => {
@@ -43,35 +43,42 @@ describe('cloudflare-a4a-config', () => {
   });
 });
 
-describe('amp-ad-network-cloudflare-impl', () => {
+describes.sandboxed('amp-ad-network-cloudflare-impl', {}, () => {
 
-  let sandbox;
   let cloudflareImpl;
   let el;
+  let doc;
+  let win;
 
   beforeEach(() => {
-    sandbox = sinon.sandbox.create();
-    el = document.createElement('amp-ad');
-    el.setAttribute('type', 'cloudflare');
-    el.setAttribute('data-cf-network', 'cloudflare');
-    el.setAttribute('src',
-        'https://firebolt.cloudflaredemo.com/a4a-ad.html');
-    sandbox.stub(AmpAdNetworkCloudflareImpl.prototype, 'getSigningServiceNames',
-        () => {
-          return ['cloudflare','cloudflare-dev'];
-        });
-    sandbox.stub(vendors, 'NETWORKS', {
-      cloudflare: {
-        base: 'https://firebolt.cloudflaredemo.com',
-      },
+    return createIframePromise().then(f => {
+      doc = f.doc;
+      win = f.win;
+      el = doc.createElement('amp-ad');
+      el.setAttribute('type', 'cloudflare');
+      el.setAttribute('data-cf-network', 'cloudflare');
+      el.setAttribute('src',
+          'https://firebolt.cloudflaredemo.com/a4a-ad.html');
+      initializeElement(el, doc);
+      sandbox.stub(
+        AmpAdNetworkCloudflareImpl.prototype, 'getSigningServiceNames',
+          () => {
+            return ['cloudflare','cloudflare-dev'];
+          });
+      sandbox.stub(vendors, 'NETWORKS', {
+        cloudflare: {
+          base: 'https://firebolt.cloudflaredemo.com',
+        },
 
-      'cf-test': {
-        base: 'https://cf-test.com',
-        src: 'https://cf-test.com/path/ad?width=SLOT_WIDTH&height=SLOT_HEIGHT',
-      },
+        'cf-test': {
+          base: 'https://cf-test.com',
+          src:
+          'https://cf-test.com/path/ad?width=SLOT_WIDTH&height=SLOT_HEIGHT',
+        },
+      });
+      doc.body.appendChild(el);
+      cloudflareImpl = new AmpAdNetworkCloudflareImpl(el);
     });
-    document.body.appendChild(el);
-    cloudflareImpl = new AmpAdNetworkCloudflareImpl(el);
   });
 
   afterEach(() => {
@@ -83,8 +90,9 @@ describe('amp-ad-network-cloudflare-impl', () => {
       expect(cloudflareImpl.isValidElement()).to.be.true;
     });
     it('should NOT be valid (impl tag name)', () => {
-      el = document.createElement('amp-ad-network-cloudflare-impl');
+      el = doc.createElement('amp-ad-network-cloudflare-impl');
       el.setAttribute('type', 'cloudflare');
+      initializeElement(el, doc);
       cloudflareImpl = new AmpAdNetworkCloudflareImpl(el);
       expect(cloudflareImpl.isValidElement()).to.be.false;
     });
@@ -113,7 +121,7 @@ describe('amp-ad-network-cloudflare-impl', () => {
       el.setAttribute('src', 'https://firebolt.cloudflaredemo.com/'
         + 'ad?width=SLOT_WIDTH&height=SLOT_HEIGHT');
       expect(cloudflareImpl.getAdUrl()).to.equal(
-          'https://firebolt.cloudflaredemo.com/_a4a/ad?width=0&height=0');
+          'https://firebolt.cloudflaredemo.com/_a4a/ad?width=200&height=50');
     });
 
     function parseQuery(query) {
@@ -151,9 +159,9 @@ describe('amp-ad-network-cloudflare-impl', () => {
       expect(url.substring(0, base.length)).to.equal(base);
       expect(parseQuery(url.substring(base.length))).to.deep.equal({
         another: 'more',
-        height: '0',
+        height: '50',
         key: 'value',
-        width: '0',
+        width: '200',
       });
     });
   });
